refactor: 만다라트 저장 로직을 커스텀 훅으로 분리
